+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 - Add internal support for 4 GB patch
 - Synchronise Steam manifest between instances
 - Add internal routing system
+- Add explanation for base game instance in setup
+- Add instance activate button
 
 ### Changed
 - Update LICENSE software author
 - Update program vendor
+- Translate missing strings in setup
 
 ### Deprecated